Refactored to hook
